// Gerdau/web/src/redux/component/types.ts
export enum ComponentActionTypes {
  FETCH_COMPONENTS = "FETCH_COMPONENTS",
  FETCH_COMPONENTS_SUCCESS = "FETCH_COMPONENTS_SUCCESS",
  FETCH_COMPONENTS_ERROR = "FETCH_COMPONENTS_ERROR",

  FETCH_SINGLE_COMPONENT = "FETCH_SINGLE_COMPONENT",
  FETCH_SINGLE_COMPONENT_SUCCESS = "FETCH_SINGLE_COMPONENT_SUCCESS",
  FETCH_SINGLE_COMPONENT_ERROR = "FETCH_SINGLE_COMPONENT_ERROR",

  FETCH_COMPONENT_EVENTS = " FETCH_COMPONENT_EVENTS",
  FETCH_COMPONENT_EVENTS_SUCCESS = " FETCH_COMPONENT_EVENTS_SUCCESS",
  FETCH_COMPONENT_EVENTS_ERROR = " FETCH_COMPONENT_EVENTS_ERROR",
}